Update deprecated operations and prepare for TensorFlow 2.0.
In particular, layers are now created using Keras objects instead of tf.layers. One side effect is that tf.get_variable() won't find the variables anymore. In TensorFlow 1.0 they can still be queried by tf.global_variables(), but this will not be true in TF 2.0. Variable names are unchanged and therefore existing checkpoints still work, with the exception of AstroFCModel, which previously used tf.contrib.layers and whose variables have been renamed from */weights to */kernel.

PiperOrigin-RevId: 262133629

+++ exoplanet-ml/astrowavenet/data/kepler_light_curves.py
@@ -36,7 +36,7 @@ def parse_example(serialized):
   # Extract values from SparseTensor objects.
   time = features["time"].values
   autoregressive_input = features["flux"].values
-  conditioning_stack = tf.to_float(features["mask"].values)
+  conditioning_stack = tf.cast(features["mask"].values, dtype=tf.float32)
   example_id = tf.cast(features["kepler_id"], dtype=tf.int32)
   return {
       "time": time,
